Shared matrix library
Adds a new `lib/matrix.ts` to handle rotation operations for matrices.
This previously was found in the day 5 code challenge, however it makes
sense to break out into a library for the usecase of day 8.

The day 5 code challenge has been updated to use the new
`rotateClockwise` matrix library function.