Shared type library
Adds `lib/type.ts` to provide functions to compare the types of all
given variables. It provides a generic `are<T>(type: string,
...variables: T[])` function, as well as a `areNumbers<T>(...variables:
T[])` function, to return whether the provided variables are of the same
specified type or numbers respectively. This is particularly useful when
provided variables that could be of multiple types, such as `string |
number`.